Fixing one test and adding some warnings

+++ test/commands/call.js
@@ -51,6 +51,7 @@ describe('apify call', () => {
     before(async function () {
         if (fs.existsSync(GLOBAL_CONFIGS_FOLDER)) {
             // Skip tests if user used CLI on local, it can break local environment!
+            console.warn(`Test was skipped as directory ${GLOBAL_CONFIGS_FOLDER} exists!`);
             this.skip();
             return;
         }
@@ -97,6 +98,7 @@ describe('apify call', () => {
     });
 
     after(async () => {
+        if (!actorId) return; // Test was skipped.
         await testUserClient.actor(actorId).delete();
         process.chdir('../');
         if (fs.existsSync(ACTOR_NAME)) await rimrafPromised(ACTOR_NAME);
